refactor: 🎨 Refactored parts of the code
Refactored Embed creation, adding a Embed class in the process. decoupled the execute commands from events.py and added them to their own file. execute now is a command group and only executable if spongiper is installed for the user, therefore locking access to these commands. added a few todos. added an Emote dataclass for easier access and changing of used emotes in the bot